Weather data updates are now animated (issue #3)
GetWeather’s onPostExecute has been completely refactored. The old
background now fades in to the new one. The logic for that has been
moved to setBackground. Our dear doge has also been animated with the
setDoge method. The status, temperature, and location text fields are
animated now too. On Gingerbread (older devices) they fade out and in
as one. On other devices they fade out and in sequentially and
(delayed) simultaneously. onPostExecute directly handles those
shenanigans.

currentTemp now always holds the API’s representation of temperature
(Kelvin) to make comparisons easier. currentlyMetric is true when we’re
displaying a metric representation of currentTemp, and false otherwise.

Need to be on the lookout for conditions where currentlyAnim is not
reset back to false. The variable is meant to prevent concurrent
updates from starting conflicting animations.