KitKat status bar and refactored duplicate styles
Had to do some weird height compensation with the sky bitmap resizing
or else the bitmap would not resize to the full width of the screen.
Not exactly certain as to why and it may not work properly for all
KitKat devices. 4px is just another magic number at this point. It
could honestly have been a bug all along (on all versions of Android)
but was hidden somehow.