Introduced the powerful WeatherUtil class
WeatherUtil is a helper class for getting up-to-date weather data from
a variety of sources. Currently these sources are OpenWeatherMap and
Yahoo. Yahoo's API tends to be more accurate but we must convert
Yahoo's weather status codes to OpenWeatherMap icon codes to display
the correct doge adjectives and images. This class is far from complete
and not yet used anywhere.

WeatherUtil will not handle the caching of weather data but it does
provide a simple WeatherData class that can be serialized for storage.

An OpenWeatherMap APP ID string is now available to be set in the build
configuration. This is optional but will be set in Google Play releases
of Weather Doge. Yahoo's weather API is public and does not provide a
means to set an API key.

Eventually WeatherUtil will be the app's primary (and only) endpoint
for retrieving new unified weather data.