Update MainActivity with WeatherUtil, Cache APIs
All the great refactoring from WeatherUtil and the Cache class is now
used by the application. There's still some work to be done to make
certain it can fail without imploding. The default weather source is
set to Yahoo but this will eventually be an option. Yahoo will most
likely replace OpenWeatherMap as the default.

This commit largely completes the initial refactoring started @ e2bf9ba