Kickass error handling for weather data
Before, error handling was relatively… poor and was pretty much
returning null if anything bad happened. Thanks to a new WeatherResult
class that wraps around WeatherData it's easy for the calling code to
see how things went. If things did not go so smoothly a good bit of
error information is available that can be reported to the user or
logged.