Rounded top corners on the widget, refactoring
The top corners of the widget are now rounded just like the bottom two.
In order to do this we have to scale the sky image on our own since
Android's RemoteViews and app widgets are so limited. Location text is
now properly aligned as well in the blue bottom bar. Some further
tweaks are still required to get everything working properly especially
with the drop shadows.

The refactored approach taken to updating multiple app widgets should
be resistant to bitmap memory leaks.