Refactored WidgetService to work on Android 4.1
Apparently cloning a RemoteViews object that holds Bitmaps is a bad
idea on Android 4.1. It appears to work just fine on nearly everything
else. To make things go more smoothly the solution was to still
generate the common bitmaps once but create a fresh RemoteViews object
each time.

Even with this fix sometimes the image generated is too small. I doubt
there’s a fix on my end since the problem appears to be with the
launcher. Perhaps provide an option for users to force the Gingerbread
compatibility image if it’s not working for them?

Sometimes getSkyBitmap will throw an exception because it is fed bogus
options data from the launcher. Not much I can do about that so I just
wrapped it in a try/catch and force it to fall back to the Gingerbread
compatibility image.

Added additional checks before recycling Bitmap objects to avoid
throwing an exception.