(Hopefully) fix empty location bug
Sometimes a location would not be available. The app would not crash
but no location would be displayed. It seemed like it could be rooted
in the geocoder logic. I previously assumed that if the Address object
is not null then it must have a locality property. This is not always
the case.

Now a WeatherData object should always have its place property set. It
will be either the search location passed in or the location retrieved
from the weather API. We can use this as a fallback in the event our
geocoder service fails.

I say hopefully because it was very difficult to pinpoint this bug and
I haven't tested enough to confirm it will not pop up again.