Add loading text on widget refresh
The location text of the widget is now changed to "Loading..." before
the refresh process. This lets the user know something is actually
happening. Perhaps something more eye-catching will be done in the
future but this is certainly good enough for now. I imagine this text
will be extra comforting for those on slower devices which may take
significantly longer to complete the refresh.