Align widget temp text to left instead of center
It would look pretty good when centered if the text below it was short
but as the text below it gets longer the temperature text would be
pushed away from the left of the screen and begin to touch doge.
Aligning to the left gives a consistent look and feel.