Vastly improved widget error handling
When the proverbial shit hits the fan, WeatherService now does a far
better job of logging what happened and giving the user a hint of what
went wrong. This small toast allows us to begin debugging before we are
given log output and alerts the user that the update process actually
finished.