Replace the widget "just now" text with the time
The time is displayed in the proper locale according to Android. Worth
noting that the time displayed is not the time the widget was refreshed
but the time that the data was retrieved from the API. This could be as
much as 10 minutes into the past after a widget refresh as the internal
cache will be invalid every 10 minutes.