Fixed an NPE occurring with poor location settings
The app could crash when location services are off. Or, more
confusingly, when location services are set to “Device only” which just
uses the GPS. Gathering an approximate location with GPS seem entirely
possible but GMS is not willing to feed us the data. Perhaps it cannot
do the approximate location fuzzing with just GPS for some reason.

Instead of crashing the app will now give the user possible solutions
via toast.