Do even more animation work in the background
setDoge has been moved into an AsyncTask reducing the amount of setup
work it needs to do on the UI thread. The GetWeather task has been
refactored to get more done in its doInBackground. A WeatherDataTask
has been created to offload some of the animation work that was done
in the GetWeather onPostExecute (UI thread!).