Switch AppWidget from AlarmManager to WorkManager
Android Oreo (8.0, API 26) applies limits to background services that
are outside of developer control. By default these limitations
generally only impact apps that target Oreo but Google Play now
requires that all new/updated apps target Oreo at minimum.

The WidgetService (IntentService) is entirely replaced with a
WidgetRefreshReceiver (BroadcastReceiver) and a WidgetWorker. The
worker can be scheduled as a periodic recurring job or as a one-off
job, which meets the needs of widget refreshes. The WorkManager API
used by the WidgetWorker is designed to work within the limitations
of Android O's background restrictions while maintaining backwards
compatibility with prior versions of Android. In fact, I believe
the AlarmManager API is used underneath for pre-Oreo devices which
means that users who aren't yet on Oreo or higher should remain
unimpacted. For users on Oreo or higher background refreshes should
now consume less resources and work more reliably since the
WorkManager should intelligently select Android's JobScheduler.

Generally, an app widget updating will always be considered a
background service except in the circumstance where the user
specifically requests an update or resizes or places down the widget.
While these special user-provoked circumstances are territory for a
foreground service, the vast majority of refreshes will be in the
background. Because of this, we'll just do every refresh as a
background task as we have before but this time using the WorkManager
API. Using a foreground service, even when the user initiates the
job, may still be undesirable since a foreground service will have a
visual notification to the user.

The WidgetRefreshReceiver allows for the new WidgetWorker to be
scheduled from the same Intent-based interface that was provided by
the WidgetService. Technically, the WidgetRefreshReceiver is only
neccesary for the widget's on-click PendingIntent refreshes, at other
times the app can simply call the WorkManager directly. But to
minimize code-churn in this large commit I'm opting to not do that
right now and instead just replacing startService calls with
sendBroadcast calls.

More on the WorkManager API:
https://developer.android.com/topic/libraries/architecture/workmanager

More on the Oreo background execution limits:
https://developer.android.com/about/versions/oreo/background