Widgets are now individually refreshed on tap
This also restores the tap-to-fresh functionality in general because
we were previously sending a ACTION_REFRESH_ALL intent only for the
WidgetRefreshReceiver to (purposefully) not handled it. Refreshing
individually means we'll be sending a (handled) ACTION_REFRESH_ONE

It was purposefully left unhandled because there is not a case where
we need to refresh all widgets through the receiver.