Disable the Yahoo! Weather API
As of January 3, 2019 the public Yahoo! Weather API has been taken
offline. This will result in errors for Weather Doge users. Until we
can reconnect to a semi-private version of the API (request pending)
or until we find an alternative provider, the app will always choose
OpenWeatherMap instead of Yahoo! even when Yahoo! is selected.